- some bugs fixed - some improvements - added generation of dmaps in util scripts - added implementation for counting using density maps

+++ utils/get_density_map_gaussian.py
@@ -6,7 +6,7 @@ from math import floor
 def get_density_map_gaussian(img, points, k_size, sigma, inverted=False):
     # points (w, h)
     [h, w] = img.shape[:2]
-    im_density = np.zeros((h, w), dtype=float)
+    im_density = np.zeros((h, w), dtype=np.float32)
 
     if len(points) == 0:
         return im_density
